pyproject.toml
[tool.poetry]
name = "scratt"
version = "0.1.0"
description = "Simple screenshot grabber with annotation capabilities."
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pycairo = "^1.20.1"
pydbus = "^0.6.0"
PyGObject = "^3.42.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
